+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Changed `view/administration/index-js.html.twig` to allow login in `sw-inactivity-login`
 * Changed login payload to store the requested redirect url
 * Changed loading of custom Admin Vue components to asynchronous loading
+* Changed SAML provider to disable the requested authn context by default
 
 **Removed**